app/requirements.txt
flask
flask-login
PyMySQL
werkzeug
flask-migrate
flask-sqlalchemy
cryptography